+++ CHANGELOG.md
@@ -36,8 +36,8 @@ from contracts import contract
 ## 3.1.0
 
 - changed port naming convention from NSEW to 1234
-- add Component.auto_rename_ports
-- add `ports_cw`, `ports_ccw` and `ports_orientation` property to ComponentReference
+- add Component.auto_rename_ports()
+- add `ports_layer` property to ComponentReference to get a map
 
 ## 3.0.2